+++ src/adobe-analytics.common.ts
@@ -12,16 +12,17 @@ export abstract class AdobeAnalyticsCommon {
         return AdobeAnalyticsCommon._instance;
     }
 
-    public abstract setContext(applicationContext: any ): void; // applicationContext type is android.content.Context
-    public abstract collectLifecycleData(activity: any, debugLogging?: boolean): void; // activity type is android.app.Activity
-    public abstract pauseCollectingLifecycleData();
-    public abstract trackState(state: string, additional: { [key: string]: any }): void;
-    public abstract trackAction(action: string, additional: { [key: string]: any }): void;
-    public abstract trackTimedActionStart(action: string, additional: { [key: string]: any }): void;
-    public abstract trackTimedActionUpdate(action: string, additional: { [key: string]: any }): void;
-    public abstract trackTimedActionEnd(action: string): void;
-    public abstract visitorAppendToURL(url: string): string;
-    public abstract trackLocation(location: any, additional: { [key: string]: any; }): void;
-    public abstract optIn(): void;
-    public abstract optOut(): void;
+  public abstract initSdk(adobeAnalyticsSettings: AdobeAnalyticsSettings, app: any): void; // app is type of globalAndroid.app.Application or UIApplication for IOS
+  public abstract collectLifecycleData(additional: { [key: string]: any }): void;
+  public abstract pauseCollectingLifecycleData(): void;
+  public abstract resumeCollectingLifecycleData(): void;
+  public abstract trackState(state: string, additional: { [key: string]: any }): void;
+  public abstract trackAction(action: string, additional: { [key: string]: any }): void;
+  public abstract optIn(): void;
+  public abstract optOut(): void;
+}
+
+export interface AdobeAnalyticsSettings {
+    environmentId: string;
+    debug: boolean;
 }