+++ src/adobe-analytics.ios.d.ts
@@ -1,8 +1,11 @@
 import { AdobeAnalyticsCommon } from './adobe-analytics.common';
+import { AdobeAnalyticsSettings } from './adobe-analytics.common';
 export declare class AdobeAnalytics extends AdobeAnalyticsCommon {
     protected static _instance: AdobeAnalyticsCommon;
-    setContext(applicationContext: any): void;
-    collectLifecycleData(activity: any, debugLogging?: boolean): void;
+    initSdk(adobeAnalyticsSettings: AdobeAnalyticsSettings, app: UIApplication): void;
+    collectLifecycleData(additional: {
+        [key: string]: any;
+    }): void;
     pauseCollectingLifecycleData(): void;
     trackState(state: string, additional: {
         [key: string]: any;
@@ -10,17 +13,7 @@ export declare class AdobeAnalytics extends AdobeAnalyticsCommon {
     trackAction(action: string, additional: {
         [key: string]: any;
     }): void;
-    trackTimedActionStart(action: string, additional: {
-        [key: string]: any;
-    }): void;
-    trackTimedActionUpdate(action: string, additional: {
-        [key: string]: any;
-    }): void;
-    trackTimedActionEnd(action: string): void;
-    visitorAppendToURL(url: string): string;
-    trackLocation(location: CLLocation, additional: {
-        [key: string]: any;
-    }): void;
     optIn(): void;
     optOut(): void;
+    resumeCollectingLifecycleData(): void;
 }