+++ src/references.d.ts
@@ -1,3 +1,13 @@
 /// <reference path="./node_modules/tns-platform-declarations/ios.d.ts" />
 /// <reference path="./node_modules/tns-platform-declarations/android.d.ts" />
-/// <reference path="./typings/ios/objc!AdobeMobileLibrary.d.ts" />
+
+/// <reference path="./platforms/ios/typings/objc!ACPCore.d.ts" />
+/// <reference path="./platforms/ios/typings/objc!ACPAnalytics.d.ts" />
+/// <reference path="./platforms/ios/typings/objc!ACPUserProfile.d.ts" />
+
+/// <reference path="./platforms/android/typings/core/android.d.ts" />
+/// <reference path="./platforms/android/typings/lifecycle/android.d.ts" />
+/// <reference path="./platforms/android/typings/signal/android.d.ts" />
+/// <reference path="./platforms/android/typings/identity/android.d.ts" />
+/// <reference path="./platforms/android/typings/analytics/android.d.ts" />
+/// <reference path="./platforms/android/typings/userprofile/android.d.ts" />